ChatHistory Panel Component is separated from Chat
