+++ code/frontend/src/components/ChatHistoryPanel/ChatHistoryPanel.module.css
@@ -1,0 +1,16 @@
+.historyContainer {
+    width: 20vw;
+    background: radial-gradient(108.78% 108.78% at 50.02% 19.78%, #FFFFFF 57.29%, #EEF6FE 100%);
+    border-radius: 8px;
+    max-height: calc(100vh - 88px);
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.14), 0px 0px 2px rgba(0, 0, 0, 0.12);
+    overflow-y: hidden;
+}
+
+.historyPanelTopRightButtons {
+    height: 48px;
+}
+
+.chatHistoryListContainer {
+    height: 100%;
+}
