+++ code/frontend/src/pages/chat/Chat.module.css
@@ -5,14 +5,6 @@
     gap: 20px
 }
 
-.historyContainer {
-    width: 20vw;
-    background: radial-gradient(108.78% 108.78% at 50.02% 19.78%, #FFFFFF 57.29%, #EEF6FE 100%);
-    border-radius: 8px;
-    max-height: calc(100vh - 88px);
-    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.14), 0px 0px 2px rgba(0, 0, 0, 0.12);
-    overflow-y: hidden;
-}
 .chatRoot {
     flex: 1;
     display: flex;
@@ -23,9 +15,6 @@
     gap: 20px;
 }
 
-.chatHistoryListContainer {
-    height: 100%;
-}
 .chatContainer {
     flex: 1;
     display: flex;
@@ -157,9 +146,6 @@
     flex-grow: 0;
 }
 
-.historyPanelTopRightButtons {
-    height: 48px;
-}
 
 .MobileChatContainer {
     @media screen and (max-width: 600px) {
